# swagger.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kasir Laundry Pro",
    "version": "1.0.0",
    "description": "API Routes and schema details of Kasir Laundry Pro Services"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/duration": {
      "get": {
        "summary": "Get all durations",
        "description": "Retrieve a list of all durations",
        "tags": [
          "Duration"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "Filter Data by name, duration, type"
          },
          {
            "in": "query",
            "name": "hasService",
            "schema": {
              "type": "string"
            },
            "description": "Filter Data by services availability"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DurationAll"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new duration",
        "description": "Create a new duration record",
        "tags": [
          "Duration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DurationRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Duration created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input"
          }
        }
      }
    },
    "/duration/{id}": {
      "put": {
        "summary": "Update a duration",
        "description": "Update an existing duration record by ID",
        "tags": [
          "Duration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the duration to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DurationRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Durasi berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input"
          },
          "404": {
            "description": "Durasi tidak ditemukan"
          }
        }
      },
      "delete": {
        "summary": "Delete a duration",
        "description": "Delete a duration record by ID",
        "tags": [
          "Duration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the duration to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Duration deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Duration deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Durasi tidak ditemukan"
          }
        }
      },
      "get": {
        "summary": "Get a duration by ID",
        "description": "Retrieve a duration by its ID",
        "tags": [
          "Duration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the duration to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurationAll"
                }
              }
            }
          },
          "404": {
            "description": "Durasi tidak ditemukan"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DurationAll": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the duration",
            "example": "1"
          },
          "name": {
            "type": "string",
            "description": "Name of the duration",
            "example": "One Day"
          },
          "duration": {
            "type": "number",
            "description": "Duration value",
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "Hari",
              "Jam"
            ],
            "description": "Type of duration (Hari or Jam)",
            "example": "Hari"
          }
        }
      },
      "DurationRequestBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the duration",
            "example": "One Day"
          },
          "duration": {
            "type": "number",
            "description": "Duration value",
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "Hari",
              "Jam"
            ],
            "description": "Type of duration (Hari or Jam)",
            "example": "Hari"
          }
        }
      },
      "DurationResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Duration created successfully"
          },
          "data": {
            "$ref": "#/components/schemas/DurationAll"
          }
        }
      },
...
    }
  },
...
}